Validate if networks exists while deleting locations
Addresses-Issue: Jira/AQUILONAQD-390

This fix is to :
- Validate if there are any networks mapped to a location still
existing while the location is being deleted.
- Modified del city code to use the common code used for delete
locations.
- Additionally added tests for organization,hub,country.

+++ lib/python2.6/aquilon/worker/commands/del_location.py
@@ -31,8 +31,9 @@
 
 from sqlalchemy.orm.exc import NoResultFound
 
-from aquilon.exceptions_ import NotFoundException
+from aquilon.exceptions_ import NotFoundException, ArgumentError
 from aquilon.worker.broker import BrokerCommand
+from aquilon.worker.dbwrappers.network import get_network_by_location
 from aquilon.aqdb.model import Location
 
 
@@ -47,5 +48,10 @@ class CommandDelLocation(BrokerCommand):
         except NoResultFound:
             raise NotFoundException("%s %s not found." %
                                     (type.capitalize(), name))
+        networks = get_network_by_location(session, dblocation)
+        if len(networks) > 0:
+            raise ArgumentError("Could not delete %s %s."
+                                " Networks were found using this location."
+                                % (type,name))
         session.delete(dblocation)
         return
